+++ lib/idkit_extension.dart
@@ -1,5 +1,6 @@
 library idkit_extension;
 
 export 'src/list_extension.dart';
+export 'src/map_extension.dart';
 export 'src/num_extension.dart';
 export 'src/string_extension.dart';