+++ test/idkit_extension_test.dart
@@ -1,7 +1,12 @@
 import 'package:idkit_extension/idkit_extension.dart';
-import 'package:idkit_extension/src/list_extension.dart';
 
 void main() {
-  String a = '6' + '0' * (-3 * -1);
-  print(a);
+  final Map<String, String> a = <String, String>{'a': '10'};
+  final Map<String, String> b = <String, String>{
+    'a': '00',
+    'b': '10',
+    'c': '20'
+  };
+  final Map<String, String> c = a.copyFrom(b);
+  print(c); // {a: 10, b: 20}
 }